requirements.txt
# Core LangChain (0.3.x ecosystem — all use langchain-core 0.3.x)
langchain>=0.3.0,<1.0.0
langchain-community>=0.3.0,<1.0.0
langchain-openai>=0.3.0,<1.0.0
langchain-chroma>=0.2.0,<1.0.0
langchain-huggingface>=0.1.0,<1.0.0    # <1.0.0 keeps it on langchain-core 0.3.x

# LangGraph (Self-RAG orchestration)
langgraph>=0.4.0

# Vector Store
chromadb>=1.0.9

# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# Document loaders
pypdf>=5.0.0
unstructured>=0.17.0
python-docx>=1.1.0

# Embeddings
sentence-transformers>=4.0.0
torch>=2.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0.0
tiktoken>=0.9.0
rich>=13.0.0
